Cleans up and Refactors the console commands with ConsoleKit
This reworks a log of the commands to be parsed with ConsoleKit and
allows easy expansion for latter on. The entire console experience is
more interactove. When fetchiong the commits, there is now a progress
bar to show some status. After the markdown is generated, it will ask
you to generate a new file.

[changelog]
added: Implemented ConsoleKit for better user interaction
removed: Removed a bunch of useless code that is not needed

ps-id: 762602DC-9977-46CB-906B-626814FA79E4
