+++ Package.swift
@@ -6,20 +6,15 @@ import PackageDescription
 let package = Package(
     name: "git-changelog",
     platforms: [
-        SupportedPlatform.macOS(SupportedPlatform.MacOSVersion.v10_13)
+        .macOS(.v10_15)
     ],
     dependencies: [
-        // Dependencies declare other packages that this package depends on.
-        // .package(url: /* package url */, from: "1.0.0"),
+         .package(url: "https://github.com/vapor/console-kit.git", from: "4.1.0"),
     ],
     targets: [
-        // Targets are the basic building blocks of a package. A target can define a module or a test suite.
-        // Targets can depend on other targets in this package, and on products in packages which this package depends on.
-        .target(
-            name: "git-changelog",
-            dependencies: []),
-        .testTarget(
-            name: "git-changelogTests",
-            dependencies: ["git-changelog"]),
+        .target(name: "git-changelog", dependencies: [
+            .product(name: "ConsoleKit", package: "console-kit")
+        ]),
+        .testTarget(name: "ChangelogTests", dependencies: ["git-changelog"]),
     ]
 )
