+++ Sources/git-changelog/Git/Bash.swift
