+++ Sources/git-changelog/Git/RunResult.swift
