+++ Sources/git-changelog/Models/Commit.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+public struct Commit {
+    public let sha: String
+    public let date: Date
+    public let body: String
+}
